MinerU/rag_test/requirements.txt
python-pptx==0.6.21
pdf2docx==0.5.6
comtypes==1.2.0
numpy==1.24.3
scikit-learn==1.3.0
requests==2.31.0
Pillow==10.0.0
python-docx==0.8.11 